Refactored code and fixed parameter splitting for Youtube url
