+++ src/config.json
@@ -1,5 +1,5 @@
 {
     "token": "NDU1MDM4NTE3OTgzMDUxODE2.XJ-7eQ.wMzWb6QJlFc14JwzepylmT9XV2Q",
     "prefix": "+!",
-    "youtube": "AIzaSyAKIvp9Tv3QsKF0kQqx-9jjy7Ys3-mIhQE"
+    "youtube": "AIzaSyBOdMMLBTqxOXGDFh6Nbk9pQmu0br1oe3Q"
 }
